pyproject.toml
[tool.poetry]
name = "portfolio-email-semder-worker"
version = "0.1.0"
description = ""
authors = ["OscarSilvaOfficial <OscarSilvaOfficial>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
kafka-python = "^2.0.2"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
